requirements.txt
numpy
tensorflow_gpu>=2.9.1
tensorflow_probability
matplotlib
scipy
keras_flops
sklearn
mat73
opencv-python
pandas
gdspy
imageio
tqdm